% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruvboxr.R
\name{deactivate_menu_theme}
\alias{deactivate_menu_theme}
\title{Deactivate gruvboxr style in file menu.}
\usage{
deactivate_menu_theme()
}
\value{
nothing.
}
\description{
Deactivate gruvboxr style in file menu.
}
